Fix mobile photo css basing image height off viewport
Would cause nasty reflow on viewport changes, which is more often than
expected. For example, scrolling up in Chrome on Android hides the
address bar and forces a reflow. The reflow loses the scroll position
and is generally disorienting. Changed to vw as I haven't encountered
any browser that changes the viewport width often.